Added proper layout configuration
Previously, the layout name was just used to load up a template html
file.  This html file would then have a specific bit removed and
replaced with content.

This commit allows for configurable layouts.  Each layout has a template
file, a snippet selector and an "insert point".  The template will be
loaded, and a snippet created using the snippet selector.  The entire
insert point will then be replaced using the snippet (and the actual
page data itself).