Extracted post code into it's own namespace
This commit extracts most of the code currently dealing with posts out
of files & render and into a separate post namespace.  The remainder of
the application hasn't been updated to support this change, so currently
things will probably be broken.

This begins the refactoring of code to support posts loaded from the
disk, and pages output to html files.